% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_distances}
\alias{compute_distances}
\title{Compute Distances Between Columns of a Matrix}
\usage{
compute_distances(matrix_input)
}
\arguments{
\item{matrix_input}{The matrix for which distances are to be calculated.}
}
\value{
A list of distance matrices and a vector of fold comparisons.
}
\description{
This function calculates various distance metrics between columns of a matrix.
}
\details{
The function calculates and returns the following distance metrics:
\itemize{
  \item \code{Euclidean}: Euclidean distance matrix.
  \item \code{Manhattan}: Manhattan distance matrix.
  \item \code{ccc}: Concordance Correlation Coefficient (CCC) matrix.
  \item \code{icc}: Intra-class Correlation (ICC) matrix.
  \item \code{CrossProd}: Cross-product matrix.
}
}
\examples{
if(interactive()){
matrixdata <- matrix(rnorm(100), ncol = 5)
result <- compute_distances(matrixdata)
}
}
