% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_precision}
\alias{compute_precision}
\title{Compute precision of sequential simulation}
\usage{
compute_precision(CY, CY_tilde)
}
\arguments{
\item{CY}{The covariance matrix of the model.}

\item{CY_tilde}{The covariance matrix of the simulated vector.}
}
\value{
The precision of sequential simulation.
}
\description{
This function computes the precision of sequential simulation by evaluating
the Frobenius norm of the difference between the covariance matrix of the model
(CY) and the covariance matrix of the simulated vector CY_tilde, normalized by
the Frobenius norm of the covariance matrix of the model (CY).
}
\examples{
CY <- matrix(c(3, 5, 7, 2, 6, 4, 0, 2, 8), nrow=3, ncol=3, byrow=TRUE)
CY_tilde <- CY
precision <- compute_precision(CY, CY_tilde)
cat("Precision of sequential simulation:", precision, "\n")
}
