% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_matrix_to_df}
\alias{convert_matrix_to_df}
\title{Convert a Matrix to a Data Frame}
\usage{
convert_matrix_to_df(mat)
}
\arguments{
\item{mat}{A matrix to be converted into a data frame.}
}
\value{
A data frame with the same content, row names, and column names as the input matrix.
If the input is not a matrix, it is returned unchanged.
}
\description{
This function converts a given matrix into a data frame while preserving its row names and column names.
If the input is not a matrix, it is returned as is.
}
