% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_gamma_initialize.R, R/optimisation.R
\name{custom_space}
\alias{custom_space}
\title{Custom search space for gamma, lambda_, and p_prime parameters}
\usage{
custom_space()

custom_space()
}
\value{
List of hyperparameter specifications

List of functions for \code{lambda_}, \code{gamma_factor}, and \code{p_prime}.
}
\description{
Defines a search space for NMF hyperparameter optimization.

Returns a list of functions generating random hyperparameter values.
}
\examples{
space <- custom_space()
}
