% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicepro-package.R
\docType{package}
\name{dicepro-package}
\alias{dicepro-package}
\title{dicepro: Semi-Supervised Deconvolution of Bulk RNA-Seq Data with Hyperparameter Optimization}
\description{
Performs semi-supervised deconvolution of bulk RNA sequencing (RNA-seq) data. Known cell-type proportions are estimated using supervised methods -- CIBERSORTx (CSx), CIBERSORT (CS), FARDEEP (Fast And Robust DEconvolution of Expression Profiles), and DCQ (Digital Cell Quantifier) -- while unknown components are inferred using non-negative matrix factorization (NMF) with limited-memory Broyden-Fletcher-Goldfarb-Shanno with bounds (L-BFGS-B) optimization. Hyperparameters are selected automatically using a Pareto-frontier-based approach with knee-point detection, allowing application when the reference signature matrix is incomplete. More details about DICEpro are available in Ba et al. (2026) \doi{ 10.1101/2026.05.XX.XXXXXX TODO (biorxiv)}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kalidouBA/dicepro}
  \item Report bugs at \url{https://github.com/kalidouBA/dicepro/issues}
}

}
\author{
\strong{Maintainer}: Kalidou BA \email{sistm.soft.maintain@gmail.com}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
}

}
