% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.custom_progress_bar}
\alias{.custom_progress_bar}
\title{Custom progress bar}
\usage{
.custom_progress_bar(total, format = "Progress [:bar] :percent", width = 60)
}
\description{
Custom progress bar
}
\keyword{internal}
