% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_experiment_paths}
\alias{.generate_experiment_paths}
\title{Generate experiment paths}
\usage{
.generate_experiment_paths(output_dir, bulkName, refName)
}
\arguments{
\item{output_dir}{Output directory}

\item{bulkName}{Bulk name}

\item{refName}{Reference name}
}
\value{
List of paths
}
\description{
Generate experiment paths
}
