% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.get_conf_from_json}
\alias{.get_conf_from_json}
\title{Load and parse configuration from JSON file}
\usage{
.get_conf_from_json(confpath)
}
\description{
Load and parse configuration from JSON file
}
\keyword{internal}
