% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.get_report_path}
\alias{.get_report_path}
\title{Get report path}
\usage{
.get_report_path(exp_dir)
}
\arguments{
\item{exp_name}{Experiment name}

\item{base_path}{Base path}
}
\value{
Report path
}
\description{
Create and return report directory path
}
\keyword{internal}
