% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.parse_config}
\alias{.parse_config}
\title{Parse configuration from JSON file}
\usage{
.parse_config(config)
}
\arguments{
\item{config}{Configuration list}
}
\value{
Parsed configuration list
}
\description{
Internal function to parse hyperparameter optimization configuration
}
\keyword{internal}
