% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.parse_hyperopt_searchspace}
\alias{.parse_hyperopt_searchspace}
\title{Parse hyperopt search space specification}
\usage{
.parse_hyperopt_searchspace(arg, specs)
}
\description{
Parse hyperopt search space specification
}
\keyword{internal}
