% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_error.R
\name{error_plot}
\alias{error_plot}
\title{Create an Error Plot}
\usage{
error_plot(error2plot)
}
\arguments{
\item{error2plot}{A data frame containing the error values to be plotted, including L1, Iterate, and value.}
}
\value{
An error plot showing the mean error values with confidence intervals.
}
\description{
This function generates an error plot to visualize error values between folds across different iterations.
}
\details{
This function calculates the mean and confidence intervals of error values between folds and creates a line plot with shaded confidence intervals. It is useful for visualizing the variation in errors across iterations and L1 categories.
}
\seealso{
Other functions for data visualization: \code{\link{ggplot}}, \code{\link{geom_line}}, \code{\link{geom_ribbon}}, \code{\link{facet_wrap}}.
}
