% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_experiment_paths}
\alias{generate_experiment_paths}
\title{Generate experiment paths}
\usage{
generate_experiment_paths(
  output_dir,
  bulkName,
  refName,
  hspaceTechniqueChoose,
  algo_select
)
}
\arguments{
\item{output_dir}{Output directory}

\item{bulkName}{Bulk name}

\item{refName}{Reference name}

\item{hspaceTechniqueChoose}{Hyperparameter space technique}

\item{algo_select}{Algorithm selection}
}
\value{
List of paths
}
\description{
Generate experiment paths
}
