% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Safe "or default" operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Object to check}

\item{b}{Default value if \code{a} is NULL}
}
\value{
\code{a} if not NULL, otherwise \code{b}
}
\description{
Returns \code{b} if \code{a} is NULL, otherwise returns \code{a}.
}
\keyword{internal}
