% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE107011.R
\docType{data}
\name{GSE107011}
\alias{GSE107011}
\title{GSE107011 dataset}
\usage{
data(GSE107011)
}
\description{
RNA-Seq profiling of 29 immune cell types and peripheral blood mononuclear cells
}
\details{
Total RNA of 29 immune cell types (from 4 individuals) and peripheral blood mononuclear cells (PBMCs, from 13 individuals)
was extracted for gene expression profiling. The 13 PBMCs samples were processed with both microarray and RNA-Seq platforms.

We performed RNA-Seq transcriptome profiling on 29 immune cell types consituting peripheral blood mononuclear cells (PBMCs)
sorted from 4 Singaporean-Chinese individuals (S4 cohort). We also performed RNA-Seq and microarray transcriptome profiling of PBMCs
from an extended cohort of 13 individuals (S13 cohort). The data was used first to characterize the transcriptomic signatures and
relationships among the 29 immune cell types. Then we explored the difference in mRNA composition in terms of transcripts proportions
and abundance. Lastly, we performed deep deconvolution for both microarray and RNA-Seq technologies.
Flow cytometry measurements of 6 cell populations:
\enumerate{
\item{Monocytes}
\item{CD4Tcells}
\item{Bcells}
\item{CD8Tcells}
\item{NK}
\item{Unknown}
}
}
\keyword{data}
