% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real_dataset.R
\docType{data}
\name{GSE107572}
\alias{GSE107572}
\title{GSE107572 dataset}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(GSE107572)
}
\description{
Blood-derived immune-cell mixtures from 9 healthy donors
}
\details{
Total RNA of 29 immune cell types (from 4 individuals) and peripheral blood mononuclear cells (PBMCs, from 13 individuals)
was extracted for gene expression profiling. The 13 PBMCs samples were processed with both microarray and RNA-Seq platforms.

We introduce quanTIseq, a method to quantify the tumor immune contexture, determined by the type and density of tumor-infiltrating
immune cells. quanTIseq is based on a novel deconvolution algorithm for RNA sequencing data that was validated with independent data sets.
Complementing the deconvolution output with image data from tissue slides enables in silico multiplexed immunodetection and provides an
efficient method for the immunophenotyping of a large number of tumor samples.
Flow cytometry measurements of 8 cell populations:
\enumerate{
\item{B.cells}
\item{T.cells.CD4}
\item{T.cells.CD8}
\item{Tregs}
\item{Neutrophils}
\item{NK.cells}
\item{Dendritic.cells}
\item{Monocytes}
}
}
\keyword{data}
