% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{hyperopt-package}
\alias{DICEPRO-package}
\alias{hyperopt-package}
\title{Hyperparameter Optimization Tools}
\description{
A complete R implementation of hyperparameter optimization tools for
non-negative matrix factorization and other machine learning tasks.
}
\details{
This package provides functions for hyperparameter optimization including:
\itemize{
\item Configuration parsing and validation
\item Search space definition and sampling
\item Optimization algorithms (random search, TPE, etc.)
\item Result visualization and reporting
}
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/yourusername/hyperopt-r}
\item Report bugs at \url{https://github.com/yourusername/hyperopt-r/issues}
}
}
\author{
\strong{Maintainer}: Kalidou BA \email{kalidou.ba@u-bordeaux.fr}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
}

}
