% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{normFrob}
\alias{normFrob}
\title{Compute Normalized Frobenius Norm of Absolute Error Matrix}
\usage{
normFrob(absolute_error_matrix, bulkData)
}
\arguments{
\item{absolute_error_matrix}{The absolute error matrix to be normalized.}

\item{bulkData}{The reference bulk data matrix.}
}
\value{
The normalized Frobenius norm of the absolute error matrix.
}
\description{
This function calculates the normalized Frobenius norm of the absolute error matrix
with respect to a reference bulk data matrix.
}
\details{
The function calculates the Frobenius norm of the absolute error matrix and
normalizes it by dividing it by the Frobenius norm of the reference bulk data matrix.
The result is a measure of the error relative to the magnitude of the reference data.
}
\examples{
if(interactive()){
error_matrix <- matrix(rnorm(100), ncol = 5)
reference_data <- matrix(rnorm(100), ncol = 5)
norm_result <- normFrob(error_matrix, reference_data)
}
}
