% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Front_pareto.R
\name{plot_kraljic}
\alias{plot_kraljic}
\title{Generate a Kraljic Pareto plot without ground-truth-based metrics}
\usage{
plot_kraljic(outputDir)
}
\arguments{
\item{outputDir}{Name of the directory saving output.}
}
\value{
A named list corresponding to the best solution on the Pareto frontier.
}
\description{
This function generates an interactive Plotly figure showing the Pareto frontier
based on Frobenius norm and constraint deviation, excluding performance indicators
calculated against ground truth (e.g., R2, NRMSE).
}
\details{
The function reads input data from a folder derived from dataset and algorithm names,
processes the data, computes the Pareto frontier using the \code{KraljicMatrix} package,
and saves an interactive HTML file. It also returns the frontier solution with the lowest
Frobenius norm and constraint deviation closest to zero (i.e., constraint closest to 1).
}
\examples{
\dontrun{
plot_kraljic(outputDir)
}


}
