% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{redirect_output_to_file}
\alias{redirect_output_to_file}
\title{Redirect output to a log file}
\usage{
redirect_output_to_file(log_file)
}
\arguments{
\item{log_file}{Character. Path to the log file.}
}
\value{
None. Output is redirected to the log file.
}
\description{
Redirects standard output and messages to a specified log file.
}
