% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computPerf}
\alias{computPerf}
\title{Compute Performance Metrics for Comparing Two Data Matrices of abundances}
\usage{
computPerf(x, y, metric)
}
\arguments{
\item{x}{A matrix of cell type proportions from the i iteration}

\item{y}{A matrix of cell type proportions from the i+1 iteration.}

\item{metric}{A metric select between adjusted R-squared (R2_adj), Root Mean Square Error Relative RMSE (RRMSE).}
}
\value{
A data frame containing performance metrics for each variable in the comparison.
}
\description{
This function computes performance metrics to compare two data matrices.
The function calculates various performance metrics, including adjusted R-squared (R2_adj),
Root Mean Square Error Relative RMSE (RRMSE).
}
\details{
The function calculates and returns the following performance metrics for each variable:
\itemize{
\item \code{R2_adj}: Adjusted R-squared.
\item \code{RRMSE}: Relative Root Mean Square Error (RRMSE).
}
}
\examples{
if(interactive()){
 x <- matrix(rnorm(50), ncol = 5)
 y <- matrix(rnorm(50), ncol = 5)
 colnames(x) <- colnames(y) <- paste0("C_", 1:5)
 result <- computPerf(x, y, "RRMSE")
}
}
