% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computPerf}
\alias{computPerf}
\title{Compute Performance Metrics for Comparing Two Data Matrices of abundances}
\usage{
computPerf(truth, estimated, it_)
}
\arguments{
\item{truth}{The true data matrix.}

\item{estimated}{The estimated data matrix to compare against the true data.}

\item{it_}{An identifier or label for the comparison.}
}
\value{
A data frame containing performance metrics for each variable in the comparison.
}
\description{
This function computes performance metrics to compare two data matrices.
The function calculates various performance metrics, including R-squared, adjusted R-squared,
Concordance Correlation Coefficient (CCC), Intra-class Correlation (ICC), Root Mean Square Error (RMSE),
Relative RMSE (RRMSE), and the name of the variable being compared.
}
\details{
The function calculates and returns the following performance metrics for each variable:
\itemize{
\item \code{R2}: R-squared.
\item \code{R2_adj}: Adjusted R-squared.
\item \code{ccc}: Concordance Correlation Coefficient (CCC).
\item \code{icc}: Intra-class Correlation (ICC).
\item \code{RRMSE}: Relative Root Mean Square Error (RRMSE).
\item \code{RMSE}: Root Mean Square Error (RMSE).
\item \code{Variable}: Name of the variable.
\item \code{Comparison}: Identifier for the comparison.
}
}
\examples{
truth_data <- matrix(rnorm(100), ncol = 5)
estimated_data <- matrix(rnorm(100), ncol = 5)
result <- computPerf(truth_data, estimated_data, "Comparison_1")
}
