% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{custom_space}
\alias{custom_space}
\title{Define custom hyperparameter search space}
\usage{
custom_space()
}
\value{
List of vectors for lambda_, gamma_factor, and p_prime.
}
\description{
Returns a restricted hyperparameter space compatible with hyperopt/research_v2.
}
