% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicepro-package.R
\docType{package}
\name{dicepro-package}
\alias{dicepro-package}
\title{dicepro: Semi-Supervised Deconvolution of Bulk RNA-Seq Data with Hyperparameter Optimization}
\description{
Performs semi-supervised deconvolution of bulk RNA-seq data. Known cell-type proportions are estimated using supervised methods (CIBERSORTx, CIBERSORT, FARDEEP, DCQ, or BayesPrism), while unknown components are inferred using non-negative matrix factorization (NMF) with L-BFGS-B optimization. Hyperparameters are selected automatically using a Pareto-frontier-based approach with knee-point detection, allowing application when the reference signature matrix is incomplete.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kalidouBA/dicepro}
  \item Report bugs at \url{https://github.com/kalidouBA/dicepro/issues}
}

}
\author{
\strong{Maintainer}: Kalidou BA \email{kalidou.ba@u-bordeaux.fr}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
}

}
