% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICEPRO-package.R
\docType{package}
\name{DICEPRO-package}
\alias{DICEPRO-package}
\title{DICEPRO: Deconvolution with Iterative Completion for Estimating cellular Proportion from RNA-seq data}
\description{
This package performs semi-supervised deconvolution of bulk RNA-seq data using the CIBERSORTx or others method. It iteratively estimates cell type proportions and refines reference data using non-negative matrix factorization (NMF) for to solve incomplete reference matrix problems.
}
\author{
\strong{Maintainer}: Kalidou BA \email{kalidou.ba@u-bordeaux.fr}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
}

}
