% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICEPRO.R
\name{DICEPRO}
\alias{DICEPRO}
\title{DICEPRO: Deconvolution and Inference of Cell Proportions with Optimization}
\usage{
DICEPRO(
  reference,
  bulk,
  methodDeconv = "CSx",
  cibersortx_email = NULL,
  cibersortx_token = NULL,
  W_prime = 0,
  p_prime = 0,
  lambda_ = 10,
  gamma_par = 100,
  N_unknownCT = 1
)
}
\arguments{
\item{reference}{A matrix of reference gene expression data, where rows are genes
and columns are cell types (cellType x gene).}

\item{bulk}{A matrix of bulk gene expression data, where rows are genes
and columns are samples (gene x sample).}

\item{methodDeconv}{A character string specifying the deconvolution method to be used.
Options include: "CSx", "DCQ", "CDSeq", "DeconRNASeq", "FARDEEP", "BayesPrism".}

\item{cibersortx_email}{An optional email address for the CIBERSORTx web-based tool.}

\item{cibersortx_token}{An optional token for the CIBERSORTx web-based tool.}

\item{W_prime}{A numeric value for the initial matrix W for unknown cell types,
default is 0.}

\item{p_prime}{A numeric value for the initial proportions of unknown cell types,
default is 0.}

\item{lambda_}{A numeric value for the regularization parameter, default is 10.}

\item{gamma_par}{A numeric value for the penalty parameter, default is 100.}

\item{N_unknownCT}{A numeric value for the number of unknown cell types, default is 1.}
}
\value{
A list containing the following components:
\item{Prediction}{The estimated cell type proportions for the bulk samples, excluding
the unknown cell type.}
\item{New_signature}{The optimized gene-cell type matrix for the cell types, including
the unknown cell type.}
}
\description{
This function performs cell type deconvolution of bulk gene expression data using
a combination of a deconvolution method and Non-Negative Matrix Factorization
(NMF) with L-BFGS-B optimization. It can integrate known cell type signatures
and estimate the proportions of unknown cell types.
}
\details{
The function performs cell type deconvolution using a specified method (e.g.,
"CSx") followed by optimization using non-negative matrix factorization (NMF).
The NMF part is performed using the \code{nmf_lbfgsb} function, where the parameters
for the optimization (e.g., \code{W_prime}, \code{p_prime}, \code{lambda_}, etc.) can be customized.
}
\examples{
\dontrun{
# Example usage of DICEPRO function
set.seed(2101)
data_simulation <- simulation(loi = "gauss", scenario = " ", bias = TRUE, nSample = 20, prop = NULL,
                              nGenes = 50, nCellsType = 30)

result <- DICEPRO(reference = data_simulation$reference,
                  bulk = data_simulation$bulk,
                  methodDeconv = "CSx",
                  W_prime = 0,
                  p_prime = 0,
                  lambda_ = 10,
                  gamma_par = 100,
                  N_unknownCT = 1)
}

}
