% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICEPRO.R
\name{DICEPRO}
\alias{DICEPRO}
\title{Semi-Supervised Deconvolution of Bulk RNA-Seq Data}
\usage{
DICEPRO(
  reference,
  bulk,
  methodDeconv = "CSx",
  metric = "RRMSE",
  cibersortx_email = NULL,
  cibersortx_token = NULL
)
}
\arguments{
\item{reference}{The reference data matrix containing cell type (column) gene expression (row) profiles.}

\item{bulk}{The bulk RNA-seq data matrix. In the columns we have the samples and in the rows the genes.}

\item{methodDeconv}{A character vector specifying the deconvolution method.
Supported values are "CSx" or "DCQ" or others. The method to use for deconvolution. Options include \code{CSx},
\code{DCQ}, \code{CDSeq}, \code{DeconRNASeq}, \code{FARDEEP} and \code{BayesPrism}.}

\item{metric}{A metric used to detect the convergence method.
Supported values are "R2_adj", "ICC", "RRMSE", "simRatio".
Choose the metric to evaluate the convergence method. Options include \code{R2_adj},
which represents Adjusted R-squared and \code{RRMSE} for Relative Root Mean Squared Error.
The convergence method will be determined based on the selected metric.}

\item{cibersortx_email}{The CIBERSORTx account email.}

\item{cibersortx_token}{The CIBERSORTx account token.}
}
\value{
A list data simulated.
}
\description{
This function performs semi-supervised deconvolution of bulk RNA-seq data using the CIBERSORTx or
Digital Cell Quantification method or others. It iteratively estimates cell type proportions and refines
reference data using non-negative matrix factorization (NMF).
}
\details{
The function calculates and returns the simulations:
\itemize{
\item \code{Prediction}: A matrix or data frame containing the estimated cell type proportions for each sample in the bulk RNA-Seq data.
\item \code{Matrix_prediction}: A matrix or data frame containing the cell type proportions for each sample in the bulk RNA-Seq data, with iteration information.
\item \code{New_signature}: The refined reference data matrix with unknown components.
\item \code{Optimal_iteration}: The optimal iteration at which the convergence criteria were met.
\item \code{performs2plot}: The variation performance between iteration.
}

This function performs deconvolution of bulk RNA-Seq data using either
the CSx or DCQ method. It first prepares the data and runs the selected method
for cell type proportion estimation. The function then calculates error norms,
performs cross-validation.
It estimates unknown components using Non-Negative Matrix Factorization (NMF)
and calculates distances between unknown components and known cell types.
The function returns the estimated cell type proportions and their associated matrices.

The function performs several steps for semi-supervised deconvolution, including:
\enumerate{
\item Running CIBERSORTx or Digital Cell Quantification or others for cell type proportion estimation.
\item Calculating and storing performance metrics.
\item Estimating unknown components using non-negative matrix factorization (NMF).
\item Computing distances between unknown components and known cell types.
\item Iteratively refining the reference data with unknown components.
}
}
\examples{
if(interactive()){
simulation <- simulation(loi = "gauss", scenario = " ", bias = TRUE,
nSample = 10, prop = NULL, nGenes = 50, nCellsType = 5)
cellTypeOut <- sample(1:ncol(simulation$reference), 2)
refDataIncomplet <- simulation$reference[,-cellTypeOut]
results <- DICEPRO(reference = refDataIncomplet, bulk = simulation$bulk, methodDeconv = "DCQ")
print(results)
}
}
\seealso{
Other functions used within deconvolution: \code{\link{running_method}}, \code{\link{nmf}}.
}
