% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICEPRO.R
\name{DICEPRO}
\alias{DICEPRO}
\title{DICEPRO: Deconvolution and Inference of Cell Proportions with Optimization}
\usage{
DICEPRO(
  reference,
  bulk,
  methodDeconv = "CSx",
  cibersortx_email = NULL,
  cibersortx_token = NULL,
  W_prime = 0,
  bulkName = "",
  refName = "",
  hp_max_evals = 100,
  N_unknownCT = 1,
  algo_select = "tpe",
  output_path = NULL
)
}
\arguments{
\item{reference}{A matrix of reference gene expression data, where rows are genes
and columns are cell types (cellType x gene).}

\item{bulk}{A matrix of bulk gene expression data, where rows are genes
and columns are samples (gene x sample).}

\item{methodDeconv}{A character string specifying the deconvolution method to be used.
Options include: "CSx", "DCQ", "CDSeq", "DeconRNASeq", "FARDEEP", "BayesPrism".}

\item{cibersortx_email}{An optional email address for the CIBERSORTx web-based tool.}

\item{cibersortx_token}{An optional token for the CIBERSORTx web-based tool.}

\item{W_prime}{A numeric value for the initial matrix W for unknown cell types,
default is 0.}

\item{bulkName}{A string indicating the name of the bulk samples dataset.}

\item{refName}{A string indicating the name of the reference dataset.}

\item{hp_max_evals}{A numeric value indicating the maximum number of evaluations for optimization.
Default is 100.}

\item{N_unknownCT}{A numeric value for the number of unknown cell types, default is 1.}

\item{algo_select}{(character) Le choix de l'algorithme d'optimisation. Options disponibles : \code{"random"} ou \code{"tpe"}. Par défaut, \code{"tpe"}.}
}
\value{
A list containing the following components:
\item{Prediction}{A matrix of the estimated cell type proportions for the bulk samples, excluding
the unknown cell type.}
\item{New_signature}{A matrix of the optimized gene-cell type matrix for the cell types, including
the unknown cell type.}
\item{W_prime}{The optimized matrix W for the unknown cell types.}
\item{P_prime}{The optimized proportions for the unknown cell types.}
\item{constraint}{The final constraint value from the optimization process.}
}
\description{
This function performs cell type deconvolution of bulk gene expression data using
a combination of a deconvolution method and Non-Negative Matrix Factorization
(NMF) with L-BFGS-B optimization. It can integrate known cell type signatures
and estimate the proportions of unknown cell types.
}
\details{
The function performs cell type deconvolution using a specified method (e.g.,
"CSx") followed by optimization using non-negative matrix factorization (NMF).
The NMF part is performed using the \code{nmf_lbfgsb} function, where the parameters
for the optimization (e.g., \code{W_prime}, \code{p_prime}, \code{lambda_}, etc.) can be customized.
The optimization process is executed via a Python script using \code{reticulate} for
interfacing with Python.
}
\examples{
\dontrun{
# Example usage of DICEPRO function
set.seed(2101)
data_simulation <- simulation(loi = "gauss", scenario = " ", bias = TRUE, nSample = 20, prop = NULL,
                              nGenes = 50, nCellsType = 30)

result <- DICEPRO(reference = data_simulation$reference,
                  bulk = data_simulation$bulk,
                  methodDeconv = "CSx",
                  W_prime = 0,
                  bulkName = "Bulk_Sample",
                  refName = "Reference_Data",
                  hp_max_evals = 100,
                  N_unknownCT = 1)
}

}
