% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{.parse_hyperopt_searchspace}
\alias{.parse_hyperopt_searchspace}
\title{Parse hyperopt search space}
\usage{
.parse_hyperopt_searchspace(arg, specs)
}
\arguments{
\item{arg}{Parameter name}

\item{specs}{Specification list}
}
\value{
Search space definition
}
\description{
Internal function to parse hyperparameter search space definitions
}
\keyword{internal}
