% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_gamma_initialize.R
\name{.sample_from_space}
\alias{.sample_from_space}
\title{Sample from search space}
\usage{
.sample_from_space(space)
}
\arguments{
\item{space}{Search space definition (list of hyperparameters with type, low, high)}
}
\value{
List of sampled parameters
}
\description{
Internal function to sample parameters from search space
}
\keyword{internal}
