% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_ref_matrix}
\alias{generate_ref_matrix}
\title{Generate a Reference Matrix}
\usage{
generate_ref_matrix(
  loi = "rpois",
  tpm = FALSE,
  bloc = FALSE,
  nGenesByCellType = 50,
  nCell = 500,
  nCellsType = 10,
  nGenes = 500,
  lam = NULL,
  pois_eps = NULL,
  corr = NULL,
  method = "Polynomial",
  sparse = FALSE,
  prob_sparse = NULL
)
}
\arguments{
\item{loi}{The law of integration to use: "rpois" for Poisson}

\item{tpm}{Logical indicating whether to convert the counts to transcripts per million (TPM).}

\item{bloc}{Logical indicating whether to introduce block sparsity.}

\item{nGenesByCellType}{Number of genes per cell type (used for block sparsity).}

\item{nCell}{Total number of cells.}

\item{nCellsType}{Number of cell types.}

\item{nGenes}{Number of genes.}

\item{lam}{The lambda parameter for the Poisson distribution (used when loi = "rpois").}

\item{pois_eps}{The epsilon parameter for the Poisson distribution (used when loi = "rpois").}

\item{corr}{The correlation matrix for generating gene expressions (used when loi = "rpois" is FALSE).}

\item{method}{The method for generating data: "Polynomial" or other.}

\item{sparse}{Logical indicating whether to introduce sparsity.}

\item{prob_sparse}{Probability of sparsity (used when sparse = TRUE).}
}
\value{
A reference matrix with cell types as columns and genes as rows.
}
\description{
This function generates a reference matrix for use in simulating single-cell RNA-seq data.
}
\examples{
if(interactive()){
set.seed(2101)
ref <- generate_ref_matrix(loi = "gauss", bias = TRUE, nGenes = 50, nCellsType = 5)
}
}
\seealso{
\code{\link{rcorrvar2}}, \code{\link{mvrnorm}}
}
