% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_ref_matrix}
\alias{generate_ref_matrix}
\title{Generate a Reference Matrix}
\usage{
generate_ref_matrix(
  loi = "rpois",
  tpm = FALSE,
  bloc = FALSE,
  nGenesByCellType = 50,
  nCell = 500,
  nCellsType = 10,
  nGenes = 500,
  lam = NULL,
  pois_eps = NULL,
  corr = NULL,
  method = "Polynomial",
  sparse = FALSE,
  prob_sparse = NULL
)
}
\arguments{
\item{loi}{The law of integration to use: "rpois" for Poisson, "splatter" for splatter-based simulation.}

\item{tpm}{Logical indicating whether to convert the counts to transcripts per million (TPM).}

\item{bloc}{Logical indicating whether to introduce block sparsity.}

\item{nGenesByCellType}{Number of genes per cell type (used for block sparsity).}

\item{nCell}{Total number of cells.}

\item{nCellsType}{Number of cell types.}

\item{nGenes}{Number of genes.}

\item{lam}{The lambda parameter for the Poisson distribution (used when loi = "rpois").}

\item{pois_eps}{The epsilon parameter for the Poisson distribution (used when loi = "rpois").}

\item{corr}{The correlation matrix for generating gene expressions (used when loi = "rpois" is FALSE).}

\item{method}{The method for generating data: "Polynomial" or other.}

\item{sparse}{Logical indicating whether to introduce sparsity.}

\item{prob_sparse}{Probability of sparsity (used when sparse = TRUE).}
}
\value{
A reference matrix with cell types as columns and genes as rows.
}
\description{
This function generates a reference matrix for use in simulating single-cell RNA-seq data.
}
\examples{
if(interactive()){

data(LM22)

set.seed(10032)
nCellsType <- 22
nGenes <- 500
nSample <- 20
lam <- runif(nCellsType, min = 5, max = 30)
pois_eps <- rep(0.000001,nCellsType)

corLm22 <- unname(cor(LM22))

sizeVal <- (nCellsType*nCellsType-nCellsType)/2
val <- corLm22[lower.tri(corLm22,diag = FALSE)]
y <- c()
for (m in 1:nCellsType\%/\%22+1) {
  y <- c(y,simcor(val,mean(val),sd(val),sample(seq(.7,.8,by=.01),1)))
}
y <- y[1:sizeVal]
corr <- matrix(1,ncol = nCellsType,nrow = nCellsType)
corr[lower.tri(corr,diag = FALSE)] <- y
corr <- t(corr)
corr[lower.tri(corr,diag = FALSE)] <- y
V <- cov(corr)
corr <- cov2cor(V)
W_gauss <- generate_ref_matrix(loi = "gauss", nCellsType = nCellsType, nGenes = nGenes, corr = corr)

}
}
\seealso{
\code{\link{zinbEstimate}}, \code{\link{zinbSimulate}}, \code{\link{rcorrvar2}},
\code{\link{mvrnorm}}
}
