% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateProp}
\alias{generateProp}
\title{This function is used for generating synthetic data representing the proportions of different cell
types in biological samples for various scenarios, which can be useful in various data analysis and
modeling tasks in the field of biology and genetics.}
\usage{
generateProp(n_cell_types, nSample, nCell, scenario = NULL)
}
\arguments{
\item{n_cell_types}{The number of different cell types.}

\item{nSample}{The number of samples to generate.}

\item{nCell}{The total number of cells in each sample.}

\item{scenario}{A scenario that determines how the proportions are generated.
It can be one of three values: "even," "uniform," or any other value.}
}
\value{
a matrix representing proportions of different cell types in samples
}
\description{
This function is used for generating synthetic data representing the proportions of different cell
types in biological samples for various scenarios, which can be useful in various data analysis and
modeling tasks in the field of biology and genetics.
}
\examples{
if(interactive()){
set.seed(10032)
prop <- generateProp(nSample = 20, n_cell_types = 22, scenario = "even")

}
}
