% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_error.R
\name{heatmap_abundances}
\alias{heatmap_abundances}
\title{Create a Heatmap of Cell Abundances}
\usage{
heatmap_abundances(res2plot)
}
\arguments{
\item{res2plot}{A data frame containing the results to be plotted, including Iteration, Cell_Type, and Abundances.}
}
\value{
A heatmap of cell abundances.
}
\description{
This function generates a heatmap to visualize cell abundances across different iterations and cell types.
}
\details{
This function uses the ggplot2 package to create a heatmap. It plots cell abundances across iterations and cell types, with colors representing the abundance levels.
}
\seealso{
Other functions for data visualization: \code{\link{ggplot}}, \code{\link{geom_raster}}.
}
