% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_error.R
\name{metric_plot}
\alias{metric_plot}
\title{Create an Error/performance Plot}
\usage{
metric_plot(perf2plot)
}
\arguments{
\item{perf2plot}{A data frame containing the metric values to be plotted, including Iterate, and metric}
}
\value{
An metric plot showing the mean metric values with confidence intervals.
}
\description{
This function generates an metric plot to visualize metric values between folds across different iterations.
}
\details{
This function calculates the mean and confidence intervals of metric values between folds and creates a line plot with shaded confidence intervals.
It is useful for visualizing the variation in metric across iterations.
}
\seealso{
Other functions for data visualization: \code{\link{ggplot}}, \code{\link{geom_line}}, \code{\link{geom_ribbon}}, \code{\link{facet_wrap}}.
}
