% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_optim.R
\name{nmf_lbfgsb}
\alias{nmf_lbfgsb}
\title{NMF with L-BFGS-B Optimization}
\usage{
nmf_lbfgsb(
  r_dataset,
  W_prime = 0,
  p_prime = 0,
  lambda_ = 10,
  gamma_par = 100,
  N_unknownCT = 1,
  con = list(maxit = 3000)
)
}
\arguments{
\item{r_dataset}{A list containing the dataset for the factorization, including:
\itemize{
\item \code{B}: A matrix of size (N_gene x N_sample), the input matrix to factorize (gene expression data).
\item \code{P_cb}: A matrix of size (N_sample x N_cellsType), cell type proportions for the samples.
\item \code{W_cb}: A matrix of size (N_gene x N_cellsType), initial matrix W (gene-cell type matrix).
}}

\item{W_prime}{Initial value for W (optional, defaults to 0). It is a scalar.}

\item{p_prime}{Initial value for p_cb (optional, defaults to 0). It is a scalar.}

\item{lambda_}{Regularization parameter (optional, defaults to 10). It is a scalar.}

\item{gamma_par}{Penalty factor for the sum of cell type proportions (optional, defaults to 100). It is a scalar.}

\item{N_unknownCT}{Number of unknown cell types to model (optional, defaults to 1). It is a scalar.}

\item{con}{A list of control parameters for the optimization process passed to \code{lbfgs3c} (optional).
\itemize{
\item \code{maxit}: Maximum number of iterations (optional, defaults to 5000).
}}
}
\value{
A list containing:
\itemize{
\item \code{w}: The optimized matrix W, a vector of length N_gene.
\item \code{H}: The optimized matrix H (cell type proportions), a matrix of size (N_sample x N_cellsType).
\item \code{loss}: The final value of the objective function (scalar).
\item \code{constraint}: The final value of the constraint term (penalty for sum of cell type proportions), a scalar.
}
}
\description{
Perform non-negative matrix factorization (NMF) using L-BFGS-B optimization. This method uses an iterative approach
to solve a regularized least-squares problem for the factorization of a matrix into non-negative matrices W and H.
It incorporates a regularization term to ensure that the solution meets certain constraints, such as cell type composition.
}
\details{
This function applies the L-BFGS-B optimization method to solve the NMF problem with a regularized objective
function. The regularization includes a penalty for the sum of cell type proportions and a scaling factor.
The optimization minimizes the objective function subject to constraints on the cell type proportions.
}
