% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R, R/optimisation.R
\name{nmf_lbfgsb_hyperOpt}
\alias{nmf_lbfgsb_hyperOpt}
\title{NMF LBFGSB Hyperparameter Optimization}
\usage{
nmf_lbfgsb_hyperOpt(
  dataset,
  W_prime = NULL,
  p_prime = NULL,
  lambda_ = 10,
  gamma = 100,
  path2save = ""
)

nmf_lbfgsb_hyperOpt(
  dataset,
  W_prime = NULL,
  p_prime = NULL,
  lambda_ = 10,
  gamma = 100,
  path2save = ""
)
}
\arguments{
\item{dataset}{List containing matrices \code{B}, \code{W}, and \code{P}.}

\item{W_prime}{Optional numeric matrix. Initial W'.}

\item{p_prime}{Optional numeric matrix. Initial P'.}

\item{lambda_}{Numeric. Regularization parameter lambda.}

\item{gamma}{Numeric. Regularization parameter gamma.}

\item{path2save}{Character. Path to save results.}
}
\value{
List. Output from \code{nmf_lbfgsb} including loss, constraints, and matrices.

List. Output from \code{nmf_lbfgsb} including loss, constraints, and matrices.
}
\description{
Calls the R function \code{nmf_lbfgsb} to perform NMF optimization with given hyperparameters.

Calls the R function \code{nmf_lbfgsb} to perform NMF optimization with given hyperparameters.
}
