% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypersearch.R
\name{objective_opt}
\alias{objective_opt}
\title{Objective function for hyperparameter optimization}
\usage{
objective_opt(
  dataset,
  config = list(),
  lambda_ = NULL,
  gamma_factor = NULL,
  gamma = NULL,
  p_prime = NULL,
  W_prime = 0
)
}
\arguments{
\item{dataset}{List. Dataset containing matrices \code{B}, \code{W}, and \code{P}.}

\item{config}{List. Configuration for the experiment, e.g., directories.}

\item{lambda_}{Numeric. Regularization parameter lambda.}

\item{gamma_factor}{Numeric. Optional multiplier for gamma.}

\item{gamma}{Numeric. Regularization parameter gamma.}

\item{p_prime}{Numeric. Optional matrix for initialization.}

\item{W_prime}{Numeric. Optional matrix for initialization.}
}
\value{
List. Contains loss, constraint, status, and current parameter values.
}
\description{
Defines the objective function to minimize for hyperparameter tuning.
}
