% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DICEPRO.R
\name{plot.DICEPRO}
\alias{plot.DICEPRO}
\title{Plot Cell Abundance Heatmap and Error Plot}
\usage{
\method{plot}{DICEPRO}(x, ...)
}
\arguments{
\item{x}{A list containing the results of both the cell abundance heatmap and error plot between folds.}

\item{...}{Additional arguments to customize the appearance of the plots.}
}
\value{
A combined plot showing the cell abundance heatmap and error plot.
}
\description{
This function combines the results of a cell abundance heatmap and an error plot for visualization.
}
\seealso{
Functions for generating individual plots: \code{\link{heatmap_abundances}}, \code{\link{metric_plot}}.
}
