% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{run_experiment}
\alias{run_experiment}
\title{Run hyperparameter optimization experiment}
\usage{
run_experiment(
  dataset,
  W_prime = 0,
  bulkName = "",
  refName = "",
  hp_max_evals = 100,
  algo_select = "random",
  output_base_dir = ".",
  hspaceTechniqueChoose = "restrictionEspace"
)
}
\arguments{
\item{dataset}{List. Dataset containing matrices \code{B}, \code{W}, and \code{P}.}

\item{bulkName}{Character. Name identifier for bulk data.}

\item{refName}{Character. Name identifier for reference data.}

\item{hp_max_evals}{Integer. Maximum number of hyperparameter evaluations.}

\item{algo_select}{Character. Algorithm selection, e.g., "random".}

\item{output_base_dir}{Character. Base directory for saving outputs.}

\item{hspaceTechniqueChoose}{Character. Hyperparameter space technique ("all" or "restrictionEspace").}
}
\value{
None. Saves JSON files and plots to disk.
}
\description{
Performs hyperparameter optimization, saves trials and results, and generates reports.
}
