% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconv_method.R
\name{running_method}
\alias{running_method}
\title{Run Deconvolution Method}
\usage{
running_method(
  bulk,
  reference,
  methodDeconv = "CSx",
  cibersortx_email,
  cibersortx_token
)
}
\arguments{
\item{bulk}{A matrix of bulk RNA-seq data where rows are genes and columns are samples.}

\item{reference}{A matrix of reference cell type gene expression profiles where rows are genes and columns are cell types.}

\item{methodDeconv}{The method to use for deconvolution. Options include "CSx", "DCQ", "DeconRNASeq", "FARDEEP", and "BayesPrism".}

\item{cibersortx_email}{The CIBERSORTx account email (used if \code{methodDeconv} is "CSx").}

\item{cibersortx_token}{The CIBERSORTx account token (used if \code{methodDeconv} is "CSx").}
}
\value{
A matrix containing the estimated cell proportions for each cell type.
}
\description{
This function performs cell deconvolution using various methods, including CSx, DCQ, DeconRNASeq, FARDEEP, etc.
}
\details{
This function allows you to run cell deconvolution using different methods. The choice of method is specified by the \code{methodDeconv} parameter.
Each method runs its corresponding deconvolution procedure and returns the estimated proportions of the cell types in the bulk data.
}
\seealso{
Documentation of the methods used for cell deconvolution: \code{\link{run_CSx}}, \code{\link{dcq}}, \code{\link{DeconRNASeq}},
\code{\link{fardeep}}.
}
