% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconv_method.R
\name{running_method}
\alias{running_method}
\title{Run Deconvolution Method}
\usage{
running_method(
  bulk,
  reference,
  methodDeconv = "CSx",
  cibersortx_email,
  cibersortx_token
)
}
\arguments{
\item{bulk}{A matrix of bulk RNA-seq data.}

\item{reference}{A matrix of reference cell type gene expression profiles.}

\item{methodDeconv}{The method to use for deconvolution. Options include "CSx", "DCQ", "CDSeq", "DeconRNASeq", "FARDEEP", and "BayesPrism".}

\item{cibersortx_email}{The CIBERSORTx account email.}

\item{cibersortx_token}{The CIBERSORTx account token.}
}
\value{
A matrix containing the estimated cell proportions for each cell type.
}
\description{
This function performs cell deconvolution using various methods.
}
\details{
This function allows you to run cell deconvolution using different methods, including CSx, DCQ, DeconRNASeq, FARDEEP...
The choice of method should be specified using the \code{methodDeconv} parameter.
}
\seealso{
Documentation of the methods used for cell deconvolution: \code{\link{run_CSx}}, \code{\link{dcq}}, \code{\link{DeconRNASeq}},
\code{\link{fardeep}}.
}
