% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulate bulk RNA-Seq Data}
\usage{
simulation(
  W = NULL,
  prop = NULL,
  nSample = 50,
  nCell = 500,
  nCellsType = 50,
  nGenes = 500,
  lam = NULL,
  pois_eps = NULL,
  corr = NULL,
  method = "Polynomial",
  scenario = NULL,
  loi = " ",
  tpm = FALSE,
  bloc = FALSE,
  nGenesByCellType = 50,
  missCellTypes = NA,
  sparse = FALSE,
  prob_sparse = 0.5,
  bias = FALSE,
  mu_bruit = 0,
  sigma_bruit = 0.025
)
}
\arguments{
\item{W}{A reference matrix, or NULL if you want to generate one.}

\item{prop}{Cell type proportions, or NULL if you want to generate them.}

\item{nSample}{Number of samples to generate.}

\item{nCell}{Total number of cells per sample.}

\item{nCellsType}{Number of cell types.}

\item{nGenes}{Number of genes.}

\item{lam}{The lambda parameter for the Poisson distribution (used when loi = "rpois").}

\item{pois_eps}{The epsilon parameter for the Poisson distribution (used when loi = "rpois").}

\item{corr}{The correlation matrix for generating gene expressions (used when loi = "rpois" is FALSE).}

\item{method}{The method for generating data: "Polynomial" or other.}

\item{scenario}{The scenario for generating cell type proportions: "even," "uniform," or other.}

\item{loi}{The law of integration to use: "rpois" for Poisson.}

\item{tpm}{Logical indicating whether to convert the counts to transcripts per million (TPM).}

\item{bloc}{Logical indicating whether to introduce block sparsity.}

\item{nGenesByCellType}{Number of genes per cell type (used for block sparsity).}

\item{missCellTypes}{Vector of cell types to exclude in the simulation.}

\item{sparse}{Logical indicating whether to introduce sparsity.}

\item{prob_sparse}{Probability of sparsity (used when sparse = TRUE).}

\item{bias}{Logical indicating whether to add bias to the generated data.}

\item{mu_bruit}{Mean value for bias (used when bias = TRUE).}

\item{sigma_bruit}{Standard deviation for bias (used when bias = TRUE).}
}
\value{
A list data simulated.
}
\description{
This function simulates bulk RNA-seq data using a reference matrix and cell type proportions.
}
\details{
The function calculates and returns the simulations:
\itemize{
\item \code{prop}: A matrix representing proportions of different cell types in samples.
\item \code{reference}: A reference matrix with cell types as columns and genes as rows.
\item \code{bulk}: A bulk RnaSeq simulated
}
}
