% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulate bulk RNA-Seq Data}
\usage{
simulation(
  W = NULL,
  prop = NULL,
  nSample = 50,
  nCell = 500,
  nCellsType = 50,
  nGenes = 500,
  lam = NULL,
  pois_eps = NULL,
  corr = NULL,
  method = "Polynomial",
  scenario = NULL,
  loi = " ",
  tpm = FALSE,
  bloc = FALSE,
  nGenesByCellType = 50,
  missCellTypes = NA,
  sparse = FALSE,
  prob_sparse = 0.5,
  bias = FALSE,
  mu_bruit = 0,
  sigma_bruit = 0.025
)
}
\arguments{
\item{W}{A reference matrix, or NULL if you want to generate one.}

\item{prop}{Cell type proportions, or NULL if you want to generate them.}

\item{nSample}{Number of samples to generate.}

\item{nCell}{Total number of cells per sample.}

\item{nCellsType}{Number of cell types.}

\item{nGenes}{Number of genes.}

\item{lam}{The lambda parameter for the Poisson distribution (used when loi = "rpois").}

\item{pois_eps}{The epsilon parameter for the Poisson distribution (used when loi = "rpois").}

\item{corr}{The correlation matrix for generating gene expressions (used when loi = "rpois" is FALSE).}

\item{method}{The method for generating data: "Polynomial" or other.}

\item{scenario}{The scenario for generating cell type proportions: "even," "uniform," or other.}

\item{loi}{The law of integration to use: "rpois" for Poisson, "splatter" for splatter-based simulation.}

\item{tpm}{Logical indicating whether to convert the counts to transcripts per million (TPM).}

\item{bloc}{Logical indicating whether to introduce block sparsity.}

\item{nGenesByCellType}{Number of genes per cell type (used for block sparsity).}

\item{missCellTypes}{Vector of cell types to exclude in the simulation.}

\item{sparse}{Logical indicating whether to introduce sparsity.}

\item{prob_sparse}{Probability of sparsity (used when sparse = TRUE).}

\item{bias}{Logical indicating whether to add bias to the generated data.}

\item{mu_bruit}{Mean value for bias (used when bias = TRUE).}

\item{sigma_bruit}{Standard deviation for bias (used when bias = TRUE).}
}
\description{
This function simulates bulk RNA-seq data using a reference matrix and cell type proportions.
}
\examples{
if(interactive()){

data(LM22)

set.seed(10032)
nCellsType <- 22
nGenes <- 500
nSample <- 20
lam <- runif(nCellsType, min = 5, max = 30)
pois_eps <- rep(0.000001,nCellsType)

corLm22 <- unname(cor(LM22))

sizeVal <- (nCellsType*nCellsType-nCellsType)/2
val <- corLm22[lower.tri(corLm22,diag = FALSE)]
y <- c()
for (m in 1:nCellsType\%/\%22+1) {
  y <- c(y,simcor(val,mean(val),sd(val),sample(seq(.7,.8,by=.01),1)))
}
y <- y[1:sizeVal]
corr <- matrix(1,ncol = nCellsType,nrow = nCellsType)
corr[lower.tri(corr,diag = FALSE)] <- y
corr <- t(corr)
corr[lower.tri(corr,diag = FALSE)] <- y
V <- cov(corr)
corr <- cov2cor(V)
resSim <- simulation(loi = "gauss", scenario = " ", bias = TRUE,
                     nSample = nSample, prop = NULL, nGenes = nGenes, nCellsType = nCellsType, corr = corr)

}
}
