% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDRnaSeq.R
\name{SSDRnaSeq}
\alias{SSDRnaSeq}
\title{Semi-Supervised Deconvolution of Bulk RNA-Seq Data}
\usage{
SSDRnaSeq(
  reference,
  bulk,
  k_folds,
  nIteration,
  cibersortx_email,
  cibersortx_token
)
}
\arguments{
\item{reference}{The reference data matrix containing cell type gene expression profiles.}

\item{bulk}{The bulk RNA-seq data matrix.}

\item{k_folds}{The number of cross-validation folds.}

\item{nIteration}{The number of iterations to perform for refining the reference data.}

\item{cibersortx_email}{The CIBERSORTx account email.}

\item{cibersortx_token}{The CIBERSORTx account token.}
}
\value{
abundances of cell types estimated.
}
\description{
This function performs semi-supervised deconvolution of bulk RNA-seq data using the CIBERSORTx method.
}
\details{
The function performs several steps for semi-supervised deconvolution, including:
\enumerate{
  \item Running CIBERSORTx for cell type proportion estimation.
  \item Calculating and storing performance metrics.
  \item Estimating unknown components using non-negative matrix factorization (NMF).
  \item Computing distances between unknown components and known cell types.
  \item Iteratively refining the reference data with unknown components.
}
}
\examples{
reference_data <- matrix(rnorm(100), ncol = 5)
bulk_data <- matrix(rnorm(100), ncol = 5)
cibersortx_email <- "your_email_example.com"
cibersortx_token <- "your_token"
results <- SSDRnaSeq(reference_data, bulk_data, 5, 50, cibersortx_email, cibersortx_token)
}
