% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_optim.R
\name{truncate2}
\alias{truncate2}
\title{Truncate numeric values to two decimal places}
\usage{
truncate2(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector truncated to two decimal places.
}
\description{
This function truncates a numeric value to exactly two decimal places
without rounding.
}
\examples{
truncate2(119.149) # Returns 119.14
truncate2(c(3.456, 7.899)) # Returns c(3.45, 7.89)

}
